Problem: application systems are big and fragile
We've been long taught to design our applications with
certain compartmentalization of the design. Here's an application
server, here's the database, here's a cache, etc. Now, when
we need to deploy and manage this application, it creates more
work for us just ot make sure everything works: deployment and
monitoring of multiple loosely connected components, configuration
management, monitoring, failovers, resource provisioning, backups, etc.

We need to make a lot of decisions, as well: what components to use,
and whether they play toggether nicely.

A similar story exists in service-oriented architectures. Services
do need to store and query their data, communicate with other services
and end-users. They also need their data replicated, backed up, etc.
They still need caching.

I am not getting any younger and this is getting tiring.

Solution: try and imagine a comprehensive platform

(I know, I know...)

This time around, there seem to be a momentum in thinking taught
your database can be *the* platform for having your application.

You're running it all the time. Your data is there. It can be
replicated, backed up (if it's any good.)

So I am going to try and build it on top of PostgreSQL. It's ubiqitous,
and it's actually pretty good. And I have some experience with its
internals.

As a first step, I am going to implement a very simplistic container
manager. What does this mean? This means if you've got a server with
PostgreSQL, and you have Docker (just Docker for the time being) installed
on it, you can let your database manage the containers.

At this time it's rather simplistic. You can create and start a container
and have it attached to that very database.

As this grows, the next order of business is to make the database
auto-restart it as conigured as well as autostarted upon database startup.

Another big thing after that is to replicate container composition functionality
so that even more complex systems can be built.

And this can get interesting. If our database server acts as an application
server as well, then we can replicate omni_containers' configuration
and all the necessary containers will be started on replicas.

More coming!