Problem: licensing terms are unclear
Solution: make it Apache 2.0

I've went back and forth on this so many times. I do like open,
free licenses just like anybody else, but I also want to earn
decent living and if my work can provide me that, that's very good.

But I also see how projects that went not-very-free (even if it's just
"can't compete with us" license) it can kind of kill the buzz and
discourage people from using or contributing it. Different story if
the license was changed after the project became popular.

So, it is what it is. Let's see if I can make something useful out
of this project, first.