Problem: seeing spontaneous crashes in omni_containers
Solution: ensure curl is initialized for every process

I've done initialization in `PG_init`, which is wrong as it's only
done during load and initialized it for that process alone. This
can mean that CURL ends up doing default initialization but we may
be freeing memory with Postgres in some situations, leading to crashes.