Problem: random failures to get id of the created container
Solution: ensure we only parse the last message in Docker's response

Docker may send multiple messages (to display progress) but we are
parsing the entire response as a whole, which is no longer a valid
JSON object but a collection of JSON objects separated by newlines.

So, instead, we iterate through all objects and use the last one.