Problem: some code is hard to test to ensure its validity
Solution: try employing formal verification to prove its properties

I am not very good at Frama-C, so take this with a grain of salt. It's
more for my curiosity at this point, but I'd love to see this evolve.

Also, be aware that it takes considerable amount of time to write anything
like this, so it should be used sparingly in some critical code paths.