Problem: polluting .pg with extensions
This makes it very difficult to work with independent builds.

Solution: deploy extensions within cmake build directory

The `deploy` and `deploy_NAME` targets are now gone.

This required a hack to make customizable PGSHAREDIR for testing purposes
but it seems to have work out alright.

Bumped up cache for .pg as it is now a patched version of Postgres.