Problem: all CI runs on Linux, even macOS one
Solution: use the correct OS for macOS

This also enables omni_containers to use DOCKER_HOST environment variable
to choose a host to connect to and DOCKER_HOST_IP to configure host's IP.

In the future, multiple contexts should be configurable in the extension's
tables instead to avoid relying on environment variables. But for the time
being, this should solve our problems with making tests on macOS work.