Problem: testing on a Linux machine while on macOS
This is often requiring a trip to CI, which is not very ergonomic.

Solution: provide a simple builder Docker container

It seems to work, however, I am having difficulty making it work
with VirtioFS, when copying files, it seems to have a problem not
finding some files (spontaneously). Copying the source directory
away from a mounted volume helps. It also appears to work fine with
gRPC FUSE, but it is *extremely* slow.

That being said, [Colima](https://github.com/abiosoft/colima)-configured
context seems to work fine, a bit slower, but bearable.

This also changes the layout of the .pg directory, adding operating
system subdirectories to allow co-existence of Postgres builds for
different platforms.