Problem: Dynpgext extensions can be useless without a loader
We can set a flag on `_Dynpgext_init` if the extension was loaded at
a startup and use it to flag to `_PG_init` that there's no Dynpgext.

However, this won't work if the extension is loaded through the following
sequence:

* CREATE EXTENSION
* SELECT omni_ext.load('extension_name') (or using other Dynpgext loader)
  This step may be a part of extension's SQL script.

At this point we can't tell if Dynpgext loader is present.

Solution: use a GUC variable to indicate Dynpgext loader presence

Open questions:

* How do we deal with multiple loaders present? Should this be possible?
* Can we expose loader function through Dynpgext interface so that
  `_PG_init` can call it so that we rely less on specific knowledge on
  how to load the extension.