Problem: extensions are provisioned altogether
This pollutes the environment of extensions that potentially creates
issues during tests or experimentation.

Solution: isolate every extension in its own share dir

As of now, this does not handle dependencies (REQUIRES) or ad-hoc
combinations of extensions for manual testing.

The reason why omni_ext tests work is because we hardcoded this dependency
in add_postgresql_extension. This whole thing needs to be improved.