Problem: exposing functionality of the database to the world
These days it typically means running a web server or an application server
of some kind that would connect to the database and interact with it
to produce an output. This creates an overhead in operations and overall
complexity.

Solution: make it possible to embed an HTTP server in Postgres

This is based on H2O, a "new generation HTTP server" with support for
HTTP/1, HTTP/2 and (experimentally) HTTP/3. I have not yet enabled support
for HTTP/3. It is also worth mentioning that even though HTTP/2 does work,
it's not parallelizing multiple streams due to the current workers architecture.
However, I do foresee this changing, just need to spend more time refining it.

Also, there's currently no support for SSL (coming) and building write AND
read handlers may be problematic at the moment due to the fact that only
top-level CTEs are allowed to write, and we're currently wrapping query as
a subquery. I have an idea for introduction of an SQL query type with functional
modifications, but I wanted to separate that work.

Bear in mind that this is just the first iteration, making it (barely)
work.

There is also a need to improve fd.c code with the ideas and notes from
https://gist.github.com/kentonv/bc7592af98c68ba2738f4436920868dc and
https://github.com/capnproto/capnproto/blob/7c8802fb9bec8818f289a44b0ec22419a845b249/c++/src/kj/async-io-unix.c++#L559

I've decided to drop Postgres 12 support as well as it's geting harder to
support it. But I am not opposed to getting it back in if there's a will
and time.