Problem: omni_httpd uses database's super user to run queries
This can be problematic when more complicated setups are used (outside of
one user for everything, basically) and lead to security issues.

Solution: lower worker to a requested & validated user

`omni_httpd.listeners` table now takes a current user (and validates that
it is current, therefore authorized) and http worker will set its role to
the correct one when needed.