Problem: omni_httpd's http_worker flow is confusing
It's hard to read and follow the logic of it

Solution: try to re-organize it a bit

The general idea here is to put things into something
closer to the order of the execution: worker process,
server initialization, request handling.

It's not yet perfect but hopefully will make it a bit easier
to navigate.