Problem: omni_httpd's example of union request "routing"
It's not great as we had to introduce priority and we're still
executing more queries than necessary trying to find the "top priority"
ones.

Solution: rewrite unions in cascading fashion
and use UNION ALL to avoid checking for duplicates.

This approach is a lot more verbose and harder to write correctly so
a helper (aggregate?) function should perhaps be provided to build
routing tables like this.