Problem: trying out Omnigres is difficult
One needs to install necessary dependencies and go through the processes
of building and running extensions. This may be okay for contributors,
but not for adopters.

Solution: provide a single Docker image one can run
with all Omnigres extensions bundled

This iteration is not perfect yet, but it's a start.