Problem: reload_configuration trigger fragility
It often works incorrectly as it is executed within the transaction,
however, it communicates with master_worker and the worker is expected
to pick up the configuration changes. If it will happen to process the
configuration change *after* the commit, then all is good. Otherwise,
the change won't be detected.

The trigger would notify the worker even if the transaction with the
change will revert. It's probably not a big deal (since no changes will
be visible to the worker) but it triggers the worker to go into the
reload mode unnecessarily.

Solution: switch to using NOTIFY/LISTEN mechanism

This is a better mechanism as it will only be executed when the
transaction succeeds and will not be duplicated (which may be important
when the configuration will grow beyond just one table.)