Problem: omni_httpd's master_worker latch
After we've started using Notify/Listen mechanism, the use-case for the
dedicated shared latch is seemingly gone. What's even worse, is that it
is relatively fragile. If the worker crashes, the latch doesn't get
disowned and therefore, on a subsequent restart, it can't be re-owned
again. This makes code really fragile.

Solution: remove the latch altogether