Problem: sharing bad file descriptor in omni_httpd
Solution: check if creating a listening socket failed

What happened is that we didn't check whether creating listening socket
was successful and tried to send `-1` to workers as a valid fd (which it
isn't)

This will now produce a warning in the log. In the future, we should
have a better alert/reporting system for improved visibility into the
system and troubleshooting.