Problem: omni_httpd.listeners table is not great
It conflates listeners and queries to execute and will not work well
when HTTPS support will be added.

Solution: normalize omni_httpd configuration

Now there are three tables (`listeners`,`sqlets` and `listeners_sqlets`)
that serve the same purpose. It gets a little noisier to work with these
but this shouldn't happen often.