Problem: changes in script files don't make it
When I change script file in an extension, and build the extension,
the new version of the script file does not get picked up.

Solution: ensure it is symbolically linked
instead of being a hard link

I've noticed that previously, only rerunning `cmake` was updating
the file.

The original intention was to have this file symlinked, but I did
not understand CREATE_LINK's default behaviour.