Problem: CTE expressions in cascading_query-created queries
If the cascading query contains a query that uses its own CTEs, this can
become problematic when it's a mutating CTE as those must be at the top
level.

Solution: move all CTEs up to the top and rename them

They will get named `<cte_name>_<sub_cte_name>`. So, if we had:

```
WITH test (SELECT 1) SELECT FROM test
```

as a query in the cascading query, test will be renamed to `test_test`
and moved to the top.

This is still not perfect if there's an actual table called `test_test`
and the query is trying to access it. This can be addressed a later
point.