Problem: omni_httpd.handlers queries can lead to runtime errors
If given queries can't be executed, they will lead to an error and
effectively HTTP 500. This is not great because (for example), a simple
typo won't be noticed until it results in a runtime error.

Solution: enforce query validity at the transaction boundary

We are using a deferred constraint trigger to verify that at the
transaction boundary, the query has to be correct.

This is not bulletproof as the query may become invalid at a later time
(say, when a relation used in it is removed), but at least it is better
than not doing any checks.

This change also brings query validation primitve to omni_sql (to be
improved at a later point)