Problem: debugging terminal on failure in CI builds
It seems to cause more trouble than the benefit at this point. Instead
of seeing a failure, I see a pending test, and have to check manually.

Solution: remove debugging terminal altogether

I want to figure out if there's a better way to do it as the next step.
But for now, let's simplify things.