Problem: omni_http config reload may still not wait for http workers
This is because of `pg_atomic_compare_exchange_u32` does:

```
The current value of *ptr will always be stored in *expected.
```

This means the semaphore we established still won't work correctly. This
is what coding late at night some time leads to.

Solution: ensure resetting `expected` on every attempt to exchange