Problem: sometimes CI can't shutdown omni_httpd
Symptom: pg_regress reporting that it can't shutdown postgres,
exit code 256

Solution: ensure we're checking if shutdown is required in tight loops

We have a few loops where we're synchronizing with other workers.
However, if the worker is being shut down, there's no point in this
synchronization and it should just stop.