Problem: having omni_httpd listeners port specified
Port may be unavailable.

Solution: allow specifying 0 to force random port selection

This also fixes a nasty bug of http_worker assuming port is Int16
(it isn't because it is signed)

I've not updated the tests to use this as this will require an
SQL-operated HTTP client as opposed to shelling out to `curl` as we do
now. I could have used a third-party extension but didn't want to deal
with their build process at the moment. We need to have our own HTTP
client anyway.