Problem: container performance is about 50% of "as is"
Noticed a major performance issue with running Omnigres from a
container. It's about twice as slow!

Solution: move it from alpine/musl to debian/glibc

I've researched this issue and it turns out musl may have problematic
components when it comes to performance in multithreaded/multiprocess
environments. Not 100% sure this is the case, but I can see with my own
eyes how changing to debian/glibs brought the performance back.