Problem: cascading_query() aggregate use and ordering
As used right now, the aggregate is not guaranteed to produce the query
in the intended order.

From Postgres documentation:

```
Ordinarily, the input rows are fed to the aggregate function in an
unspecified order. In many cases this does not matter; for example, min
produces the same result no matter what order it receives the inputs in.
However, some aggregate functions (such as array_agg and string_agg)
produce results that depend on the ordering of the input rows. When
using such an aggregate, the optional order_by_clause can be used to
specify the desired ordering.
```

Solution: use `ORDER BY` clause inside the aggregator.