Problem: ocassional assert failures in libh2o
Solution: avoid incorrect disposal of listener contexts

The assertion failures are happening because some of the memory
pointed to has been reclaimed.

Turns out, there is a bug in http_worker logic: it'll try to get new fds
from the master_worker every time the event loop exits, without checking
whether a worker reload flag has changed.

Since the currently used socket fd hasn't been closed, the socket
received by http_worker has a different numeric value and the comparison
in the logic leads to a decision to dispose of the listener context for
the (wrongly assumed to be removed) fd.

However, there are still requests in flight, on a live socket and this
rips out an allocation that is expected to be valid.

By ensuring that socket reload is done only when a reload is requested
by master_worker, we, at the very least, minimize chances of this
happening.

What needs to happen here is that the comparison logic needs to be
changed to something more robust.

Also, I have a suspicion that because STC is using current memory
context in master_worker, it may eventually be [partially] reclaimed when
the context created by SPI_connect is reclaimed, resulting in undefined
behaviour. This is not good, but entirely fixable. Just need to make a
decision whether using Postgres' allocators in STC was a sound decision
at all. Need to do further analysis.

For the time being, this patch is seeminly preventing the assertion
failures.