Problem: updating omni_httpd.listeners port
When a port is specified as 0, omni_httpd will update it with the actual
port. However, this means that on the restart, it'll keep using that
port that was selected once.

The intention of using zero is to say, "pick one." And restarts should
honour this.

Proposed solution: add effective_port column

This way omni_httpd can use it always update to set the selected port.

While at it, changed relevant SQL code in C files to be lowercase to
match the style elsewhere.

Closes #111