Problem: omni_httpd workers are blocking
Every HTTP worker is currently going busy when getting a new connection and a request.

This works (and having more than one worker ensures there are more
workers to accept new connections), however, it negatively impacts
HTTP/2 multiplexing. Once the worker is processing the request, it will
not be processing other requests.

Solution: run HTTP processing event loop in a separate thread
and use the primary thread for maintenance operations and executing
Postgres handlers.

In addition, if there are requests in flight (defined as Postgres
executing something), and the protocol is 2 or above, proxy the request
back. It's not perfect as at times it is faster to wait than reprocess
the query elsewhere, but we'd need a more sophisticated mechanism to
handle this. Or maybe more tuning knobs.

Resolves #40