Problem: omni_httpd tests hanging in Release mode
Solution: move SIGTERM/SIGUSR2 handling to the main thread

My best understanding of this issue is that since we were installing
signal handler in a thread, sometimes it was done at a wrong time and
therefore some configuration changes (affected by SIGUSR2) or
terminations may not have been caught. I am not 100% sure about this,
but it appears to work now.