Problem: sum types can't be changed
This means that if we want to add a variant, we have to create a new
type and migrate to it.

Solution: allow amending sum types

For variably sized sum types, adding any new variants is permitted. For
fixed sized ones, only variants of the same size or smaller are
accepted. The reason for this is that if it was larger, then we'd have
to change the size of the type and Postgres would attempt to use this
size to read existing values which is clearly wrong.