Problem: sequence counters do not get replicated
From Postgres documentation on [restrictions of logical
replication](https://www.postgresql.org/docs/current/logical-replication-restrictions.html)

```
Sequence data is not replicated. The data in serial or identity columns
backed by sequences will of course be replicated as part of the table,
but the sequence itself would still show the start value on the
subscriber. If the subscriber is used as a read-only database,
then this should typically not be a problem. If, however, some
kind of switchover or failover to the subscriber database is
intended, then the sequences would need to be updated to the
latest values, either by copying the current data from the
publisher (perhaps using pg_dump) or by determining a
sufficiently high value from the tables themselves.
```

Solution: prefix local counter values with a node id / system identifier

This way each replica maintains its own counters but IDs are prefixed
with the node's identifier.

This is somewhat less efficient as it stores and allocates a bit more
data.

The implementation is not perfect and is not grealy optimized yet.
At the very least, there's potential for packing smaller pairs that fit
into `Datum` into passing by values instead of pointers.

As part of this commit, I've added Jinja library wrapper that is used to
(manually) generate heaps of repetitive code. This is not automated yet
as doing this with CMake was too much of a rabbit hole for the time
being. In the initial implementation, I was doing all the initialization
using C API (to accommodate the repetitivesness), but it was getting
large and not very easy to follow, so I had to go back to defining it in
C.