Problem: some responses have garbage in headers
Solution: ensure we actually allocate them on h2o pool

The way headers were passed, they were simply pointers to
Postgres-allocated memory. It worked before when we sent response
inline, in one process. But we now send them in a different I/O thread
and that was causing problems.

I've made sending body follow the same pattern for consistency.

It looks like we're now able to get higher throughput with this. Not
sure how to explain this, though. In my limited benchmarking, up to
two times faster than before!

I wonder if we can get rid of extra allocations/copies here in the
future. For example, by de-parenting the memory context and sweeping
leaked contexts at a later time?