Problem: omni_httpd seems to hang
Given sufficient amount of requests, the CPU starts spinning
and no new requests are served, at least not for a while.

Solution: ensure we don't let evloop poll it again

By simply returning from on_accept, evloop schedules it again to read,
in perpetuity. We disable accepting until the number of requests in flight
comes down to zero.

Importantly, we should decrement the number on the right thread
(event_loop)