Problem: using role_name in omni_httpd.handlers
This refers to roles by name, which are resolved to OIDs during
configuration reload.

This is potentially problematic as it means the role can be unavailable
at the time of the configuration reload.

It can be fragile if something is to happen to the search path in the
future.

Solution: make `role_name` become `role` of type `regrole`

It's not *substantially* better but it feels like it's more of the right
thing to do.

This can enable us to use `pg_shdepend` to ensure we can't drop the
roles that are used.