Problem: using pg_regress
The idea behind pg_regress is great: write out your psql session
and check if it is going to be the same when you run your code again
(perhaps after changing it.)

However, these scripts often quickly devolve into something that is
difficult to manage as it's a lot of plain text that has to be separated
and organized in some way.

Solution: organize tests in YAML files

I've built an early version of pg_yregress that takes YAML-based
description of tests, runs through them and outputs the resulting
YAML, returning 1 if they don't match.

Tools like [dyff](https://github.com/homeport/dyff) can be then used to
spot the differences. There's also
[yaml-merge](https://github.com/alexlafroscia/yaml-merge) but I haven't
tried it yet.

The tool is currently rather crude and is not documented, but I am going
to evolve it quickly as I need a better regression testing story.

It's not currently used for any extensions, but it will be soon.