Problem: multiple test cases in pg_yregress
Having to write a lot of cases or tons of data points to make a test
is at times tedious.

Solution: implicitly support Inja templates to generate tests
(or anything, really)

The trade-off in this approach is that you will not be able to use
results of `pg_yregress` to overwrite the original file when you're
updating tests. But that has always been a bit of an issue as it changed
formatting and seemingly eliminated comments.