Problem: pg_yregress under-reports tests
This is happening with its own test suite.

Solution: ensure errored transactions are restarted
and pg_yregress counts skipped tests.

If steps are skipped because the sequence of steps
has failed, count them as skipped.

However, if the transaction has failed expectably (an expected error),
roll it back and start it anew so that remaining tests can be used.
This helped finding a test that should have failed but didn't.