Problem: Work In Progress tests in pg_yregress
Sometimes tests are still being developed or depend on some external
future events to happen before becoming valid.

This is not properly supported by pg_yregress. One can use `skip` but
it'll not execute them at all.

Solution: designate a `todo` directive similar to `skip`

It'll run the test but use `TODO` reporting directive regardless of the
outcome, which can be used by TAP consumers to report the fact that
there are tests that are WIP.