Problem: omni_httpd's role.yml test is flaky
At times, HTTP connections are refused

Solution: ensure we wait for the initial omni_httpd load

From https://github.com/omnigres/omnigres/issues/186#issuecomment-1570508233:

```
I think this is what is happening:

When omni_httpd is started with no_init (no default listener), upon
startup, the master worker reads the configuration, finds no listeners,
and proceeds to insert a record into configuration_reloads, which
indicates that a reload has happened. And in a sense, it has, however,
it is a "nil" configuration. We could make it conditional, saying
there won't be a reload notification if there are no listeners (or the
    listener configuration hasn't changed). However, this is
problematic:

We don't check the handlers and if their configuration changed (this is
currently the responsibility of http workers). We could, however,
try and do this.  It will have negative impact on interface
expectations. Let's say we issue an update to the configuration but
it doesn't change anything or deletes all listeners. Shouldn't there
 be a notification? Of course, there should be!

So, perhaps, with this in mind, the test scripts need to ensure they
wait for a configuration reload of the initial configuration. Or wait
for two reloads. This is not ideal, so I am still open to other ideas.
```

It has seemingly made the test less flaky but didn't seem to eliminate
it altogether.