Problem: omni_httpd master worker coordination
It's a bit complicated: we have one transaction in which we start the
listener (picking a port, if necessary), then we commit it and notify
the HTTP workers that they can reload. We commit so that they can
read the effective port so that they can find the socket by address and
port.

The HTTP workers then notify the master worker that they are ready to
receive a list of sockets.

After, this we have a second transaction that signals that we're done
reloading the configuration so that user code can rely on that for
testing and other purposes.

Solution: change the socket lookup method

By sorting listeners by ID (ascending), we always get the same order of
IDs and, therefore, can use the vector of received sockets to get them by
index instead.

In order to guarantee that the order won't change, master worker locks
the `listeners` table before it reads it and other transactions can
only read from it but not modify it.

This simplifies the socket discovery and gives us a single transaction
in the master worker.