Problem: occasional segfault trying to pthread_mutex_lock
Solution: register event_loop's receiver before creating a thread

This one was hard to trace, as core dumps were only showing one frame,
but it did point to the fact the mutex was attempted to be locked at
0x10, which happens to be the offset of the mutex in h2o's
`h2o_multithread_queue_t`.

I am still not 100% confident what exactly made it behave like this
(making it look like `queue` is `NULL` in certain cases), but most
likely it is this:

* We initialize the queue (`setup_server` in `http_worker`)
* We start the `event_loop` thread, which [at some point] registers a
  receiver
* `http_server` starts receiving traffic, and uses one of the
  `h2o_queue_send` functions
* `h2o_queue_send` family uses `event_loop_receiver` to send messages
  back to the event loop.

Being called from a different thread, the receiver in this thread may
not be having a valid value, leading to a `NULL` `queue` in it.

Hence the bug.

The fix is not perfect as it is still messy and is in need in minor
rearrangements to make this more straightforward. But it seems to fix
the problem.