Problem: dev env and deployment workflows aren't great
Getting files from the development environment and repositories is not
straightforward. This results in less than great crutches to get the
information in.

Solution: provide a unified Virtual File System API

This kickstarts the omni_vfs extension. At this time, it only features a
limited version of the `local_fs` filesystem intended to be used in
development environments.