Problem: developing for databases is not always fun
A particularly annoying aspect of it is that functions and policies end
up being in migrations and it is hard to track changes across them as
they end up being in endless migrations.

Solution: devise an object-tracking mechanism for these

If we track such objects (functions, policies) created, we can easily
change them during re-deployments (whether those happen in production or
locally in development environment).

To this, I've created a new extension `omni_schema` that currently does
just that is implemented in pure PL/pgSQL for the time being.

I also want it to support migrations (one of the next steps) and expand
its ability to re-create other objects from a single source of truth.
This will allow us be able to track changes to such objects as tables,
types and others and migrations will be rather addressing the issue
of data migrations and corner cases.

I also want it to support discovering sources for functions implemented
in other languages (like Python, Perl, Rust, etc.) and including those
into the process. Because, honestly, who likes writing code inside of
`create function`'s `$$ $$`?