Problem: omni_vfs.local_fs() transfer of the result
It doesn't actually happen and manifests in other pieces of code getting
garbage instead of proper data.

Solution: ensure we capture the result of `SPI_transferDatum`

Also, remove the unnecessary manipulations of the memory context as
these are done by `SPI_transferDatum`. They are only necessary if we use
`transferDatum`.