Problem: handling database schema changes
In a lot of cases, it is difficult or impossible to "reload" the
database with a DDL like we do for functions or policies, as either the
data needs special handling or the way the transformation is done, can't
be achieved by mechanical diffing and changing the object.

However, we still need to hangle changes to the schema.

Solution: devise a simplistic migration facility

It will run a previously-unknown migrations in order. Nothing fancy so
far.

This is still not great for tracking the overall structure of objects
derived this way.

Two things that can improve this:

1. Dumping DDL back so that we can check it into the development
   repository. This is what Ruby on Rails does, for example. However,
   there's a bit of a problem here. For this to work, `pg_dump` is
   required and it is a separate process, so it's definitely somewhat of
   the scope of this extension. However, since this is to be executed by
   the development environment tooling (and never in deployed systems),
   may be that's ok.

2. Automatic diffing of relations. In a number of straightforward
   scenarios we can, in fact, allow to store relation definitions as
   they are and generate alteration statements. A good question here is
   whether they should be generated in runtime or build-time to be
   inserted into "migrations" directory and be checked into the
   repository. If it is done in runtime, we can't fully vet what's going
   on, and that's scary.