Problem: omni_schema functions always require a sub-path
However, this is annoying as often it would be just an empty string.

Solution: make it an optional parameter

By default, it'll just use the root of the given file system.

That there may be limitations for creating new file systems, and that's
why sub-path can still be specified if there's only a limited number of
file systems that can be accessed.