Problem: our build process only uses self-managed Postgres builds
This is problematic when we want to be able to build exensions against
Postgres installed using system-wide package manager or any other
method.

For example, this would be useful for building extensions for specific
Linux distros.

Solution: allow this by taking `PG_CONFIG` variable to specify `pg_config`

The command would now look like:

```shell
cmake -DPG_CONFIG=/path/to/pg_config # for example, $(which pg_config)
```

We won't be able to run psql or test targets yet because of how we ship
extensions using patched Postgres. Same applies to Frama-C tests. But I
think that this is okay for the time being. If you want to develop and
test, self-managed Postgres builds are okay. It would be great to
improve this in the future, but not a huge priority at the moment.