Problem: pg_yregress uses Postgres notation for booleans
That is, it will put `t` or `f` into results. However, this is
inconsistent with the rest of the YAML file where booleans are denoted
standard YAML scalars for booleans.

Solution: use `true` and `false` values for booleans

This solution is not yet perfect because the comparison will not work
correctly for different variations of the boolean scalars in YAML.
Fixing this will require implementing custom comparator and this hinges
on being able to (ideally) reuse the standard comparison algorithms with
boolean comparison on top of it. I've previously sent a PR:
https://github.com/pantoniou/libfyaml/pull/85 but it wasn't merged yet.