Problem: extensions can't be built individually
This is because of their assumption that they are built in-tree.
However, this would be very useful for packaging to avoid that.

Solution: fix build process to accommodate this need

I am absolutely not loving the additional mess I put into
CMakeLists.txts here to avoid double-inclusion of subdirectories and
the additional complexity around building the extensions. This is
definitely technical debt, but at least it gets us to shipping
extensions sooner. It must be paid off (doesn't everybody intend to when
they say this?)

Also, test that this works in CI. The reason why I didn't parallelize
the invocation of the target here is because right now there's no
synchronization for building Postgres in `.pg` and all these tasks will
race to happily do so. Once there's synchronization in place, this can
be parallelized.