Problem: repeated subdirectory exclusion has a flaw
It manifests in not being able to successfully run `cmake
/path/to/omnigres` the second time, it'll complain about not being able
to find one of the components.

Solution: ensure guards don't survive between the calls

To imitate global variables for inclusion guards, I've used cached
variables. However, they survive across the invocations of `cmake` and
this means the guard off the legitimate inclusions of the directories
they are guarding.

I've made them not do that by using environment variables instead.