Problem: lack of distinct options to process data
Your choices are either SQL with all its expressive power,
potential parallelization, optimizations, etc. or a few imperative or
functional languages, starting from PL/pgSQL.

Other approaches are not quite available.

Solution: provide a proof-of-concept integration of Prolog

It uses SWI-Prolog as the implementation of choice. It's not the fastest
one, or ISO-compliant. But it does integrate relatively well and packs a
lot of features. It's actively developed and has a good community.
Being a relative newcomer to Prolog, I can't tell if this is a perfect
decision long-term, but it does make sense now.

Why Prolog?

As I was working on Postgres package management, I ended up choosing
Prolog as the language for describing constraints and rules for building
packages in varieties of configurations and settings. And since I was
integrating it with C codebase, I thought: "hey, maybe this is a good
candidate for a language in Postgres!". It certainly allows to process
data from the database in a different, unique, and valuable way.

This current implementation is incomplete but shows the general idea and
even implements a sandboxed (trusted) version of the language using
`library(sandbox)`.