Problem: YAML output of pg_yregress
When trying to find an error, it's not very helpful as it requires
comparing with the original document. And it can't be compared properly
at the moment due to the use of references and some other artifacts.

Solution: make TAP the primary output

And make the original output optional. Make sure postgres logs are
collected and not printed out together with the TAP output.