Problem: having no error doesn't clear error
pg_yregress won't clear `error` and thus fail if error
did not occur but was specified.

Solution: ensure `error` is cleared

Turns out we had some places doing this all wrong and not testing
correctly. The most prominent example if testing a trigger that only
fires at the transaction boundary. Those tests never checked what they
were supposed to test because we were not trying to commit.

I've added some (admittedly convoluted) logic to try and commit the
transaction before checking the results so that we can rely either on
query failure or transaction failure for the error or on the success of
the query for the results.